Slight modifications
- No longer saving ALL export files. When a new export is requested, we
clean out our folder before saving. We could remove this and add some
sort of ‘backup’ feature that lists all past exports on the ‘Export’
screen with download links.
- If we don’t have direct access to write files to the server, we’ll
try to download the file via javascript. Chrome downloads the file,
Safari displays the content in a window, Firefox gives you a prompt to
open or download.