Added an AJAX call to flush the rewrite rules on options save
I forgot to account for when a user would change the `staff-member`
slug on the plugin options page. Changing the slug would result in 404s
but now, after detecting a settings change, we flush the rewrite rules
via AJAX. I did it this way because of the way the options page is
written. It just updates the values after the page refreshes, so
flushing them on the page is too late.