Add the `sslp_query_args` filter
Adds the `sslp_query_args` filter that filters the args used in `WP_Query` by the `[simple-staff-list]` shortcode. The second parameter is `$context` and describes the context in which the filter is being run. I added this parameter in case there are other scenarios where we're querying staff members but _not_ from the shortcode i.e. a future Gutenberg block.

If an array is not returned, I reset `$args` to the unfiltered value. I then assign the `post_type` value so we're always going to be querying the `staff-member` post type.