Better Export for Custom Fields
I have added custom fields to SSLP via a custom plugin. For some reason, when I export the data the custom fields are out of order. (For example, I would have phone numbers showing in the Bio column on the csv.) If you write the custom data by directly referencing the key name (probably not saying that right) rather than looping through each key -> value pair, I can force the export to retain the correct order.

Not sure if that is useful to anyone else, but it works for me. Until there is support for custom fields in this plug in.