Renamed single-staff-member block to single-staff-member-legacy
I'd like to offer a better way to template out the staff member loops. I'm marking this block as "legacy" because we're using the shortcode to generate the output. I'll be building "new" blocks that will have a few different layout options and the ability to hook into them and add additional layouts.